requirements.txt
asgiref==3.3.1
Django==3.1.7
django-rest-framework==0.1.0
djangorestframework==3.12.2
mysql-connector-python==8.0.23
mysqlclient==2.0.3
protobuf==3.15.3
PyJWT==2.0.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
